refactor(desktop): builder registries + wizard extraction in hosts/_common.py
Replaces the hand-written if/elif chains in build_storage() and
build_action_chains() with _STORAGE_BUILDERS / _ACTION_EXTENSION_BUILDERS
registries (dict/list of (predicate, builder) pairs), so adding a new
storage backend or pre/post-like action extension is a registration
(one _build_* function + one registry entry) instead of a new branch.

Also extracts the interactive --setup wizard out of _common.py into its
own hosts/_setup.py module, since it's a distinct concern (interactive
I/O) from _common's config/storage/action-chain wiring. _setup.py reaches
back into _common via qualified attribute lookups (_common.load_config(),
_common.make_provider(), etc.) rather than `from ._common import X`, so
existing tests' monkeypatch.setattr(_common, ...) calls keep working
unchanged.

Trigger construction (make_tray_hotkey_trigger / make_one_shot_cli_trigger)
is intentionally left out of the registry — it's host-specific, not
cfg-dispatched within _common.py.

Closes #58
